+++ packages/select/docs/Select/Select.stories.tsx
@@ -1,14 +1,13 @@
 import React from "react"
-import { useSelectState } from "@react-stately/select"
+import { Node } from "@react-types/shared"
 import { ComponentMeta, ComponentStory, Story } from "@storybook/react"
 import { withDesign } from "storybook-addon-designs"
-import { StoryWrapper } from "../../../storybook/components/StoryWrapper"
-import { CATEGORIES, SUB_CATEGORIES } from "../../../storybook/constants"
-import { figmaEmbed } from "../../../storybook/helpers"
-import { Select, getSelectChildren } from "../src/Select/Select"
-import overlayStyles from "../src/Select/components/Overlay/Overlay.module.scss"
-import { SelectContext } from "../src/Select/context/SelectContext"
-import { singleMockItems } from "./MockData"
+import { StoryWrapper } from "../../../../storybook/components/StoryWrapper"
+import { CATEGORIES, SUB_CATEGORIES } from "../../../../storybook/constants"
+import { figmaEmbed } from "../../../../storybook/helpers"
+import { Select } from "../../src/Select/Select"
+import { SingleItemType } from "../../src/types"
+import { singleMockItems } from "../MockData"
 
 export default {
   title: `${CATEGORIES.components}/${SUB_CATEGORIES.select}/Select`,
@@ -60,47 +59,6 @@ DefaultStory.args = {
 DefaultStory.parameters = {
   chromatic: { disable: false },
   docs: { source: { type: "code" } },
-}
-
-type MockListBoxProps = {
-  optionClassName?: string
-  selectedKey?: React.Key | null
-  isFullWidth?: boolean
-  disabledValues?: React.Key[]
-}
-
-const MockListBox = ({
-  optionClassName,
-  selectedKey,
-  isFullWidth,
-  disabledValues,
-}: MockListBoxProps): JSX.Element => {
-  const mockState = useSelectState({
-    selectedKey: selectedKey ?? undefined,
-    items: singleMockItems,
-    children: getSelectChildren,
-    disabledKeys: disabledValues,
-  })
-  return (
-    <SelectContext.Provider value={{ state: mockState }}>
-      <div
-        className={overlayStyles.menuPopup}
-        style={{ position: "relative", width: !isFullWidth ? "180px" : "100%" }}
-      >
-        <Select.ListBox menuProps={{}}>
-          {Array.from(mockState.collection).map(item => (
-            <Select.Option
-              key={item.key}
-              item={item}
-              classNameOverride={
-                item.key === "id-sre" ? `${optionClassName}` : undefined
-              }
-            />
-          ))}
-        </Select.ListBox>
-      </div>
-    </SelectContext.Provider>
-  )
 }
 
 const StickerSheetTemplate: Story<{ isReversed: boolean }> = ({
@@ -219,24 +177,28 @@ const StickerSheetTemplate: Story<{ isReversed: boolean }> = ({
       </StoryWrapper.Row>
     </StoryWrapper>
 
-    <div style={{ height: "550px", marginTop: "12rem" }}>
-      <StoryWrapper isReversed={isReversed}>
-        <StoryWrapper.RowHeader
-          headings={["Base", "Selected", "Hover", "Focus", "Disabled"]}
-        />
-        <StoryWrapper.Row rowTitle="Dropdown">
-          <MockListBox />
-          <MockListBox selectedKey="id-sre" />
-          <MockListBox optionClassName="story__option-hover" />
-          <MockListBox optionClassName="story__option-focus" />
-          <MockListBox disabledValues={["id-sre"]} />
-        </StoryWrapper.Row>
-      </StoryWrapper>
-    </div>
-
     <StoryWrapper isReversed={isReversed}>
-      <StoryWrapper.Row rowTitle="Dropdown Fullwidth">
-        <MockListBox isFullWidth />
+      <StoryWrapper.RowHeader headings={["Truncated"]} />
+      <StoryWrapper.Row rowTitle="Long Texts">
+        <Select
+          id="select-long-truncated"
+          label="label"
+          items={[
+            { label: "Dev-ops", value: "id-devops" },
+            { label: "Others", value: "id-others" },
+            {
+              label:
+                "Super long option where the container is fixed width and the selected option goes multiline",
+              value: "id-long",
+            },
+            {
+              label: "Metallblasinstrumentenbauermeisterbrief",
+              value: "id-long-word",
+            },
+          ]}
+          description="This is a description"
+          selectedKey={"id-long"}
+        />
       </StoryWrapper.Row>
     </StoryWrapper>
   </>
@@ -245,6 +207,104 @@ const StickerSheetTemplate: Story<{ isReversed: boolean }> = ({
 export const StickerSheetDefault = StickerSheetTemplate.bind({})
 StickerSheetDefault.storyName = "Sticker Sheet (Default)"
 StickerSheetDefault.parameters = {
+  chromatic: { disable: false },
+  controls: { disable: true },
+}
+
+const DropdownSheet: Story<{ isReversed: boolean }> = ({ isReversed }) => (
+  <>
+    <div style={{ marginBottom: "28rem" }}>
+      <StoryWrapper isReversed={isReversed}>
+        <StoryWrapper.RowHeader
+          headings={["Base", "Selected", "Hover", "Focus", "Disabled"]}
+        />
+        <StoryWrapper.Row rowTitle="Dropdown">
+          <Select
+            id="select-dropdown-base"
+            label="label"
+            items={singleMockItems}
+            placeholder="Placeholder"
+            isOpen
+          />
+          <Select
+            id="select-dropdown-selected"
+            label="label"
+            items={singleMockItems}
+            placeholder="Placeholder"
+            selectedKey="id-sre"
+            isOpen
+          />
+          <Select
+            id="select-dropdown-hover"
+            label="label"
+            items={singleMockItems}
+            placeholder="Placeholder"
+            isOpen
+          >
+            {(optionsProps): JSX.Element[] =>
+              optionsProps.items.map((item: Node<SingleItemType>) => (
+                <Select.Option
+                  {...optionsProps}
+                  key={item.key}
+                  item={item}
+                  classNameOverride={
+                    item.key === "id-sre" ? "story__option-hover" : undefined
+                  }
+                />
+              ))
+            }
+          </Select>
+          <Select
+            id="select-dropdown-focus"
+            label="label"
+            items={singleMockItems}
+            placeholder="Placeholder"
+            isOpen
+          >
+            {(optionsProps): JSX.Element[] =>
+              optionsProps.items.map((item: Node<SingleItemType>) => (
+                <Select.Option
+                  {...optionsProps}
+                  key={item.key}
+                  item={item}
+                  classNameOverride={
+                    item.key === "id-sre" ? "story__option-focus" : undefined
+                  }
+                />
+              ))
+            }
+          </Select>
+          <Select
+            id="select-dropdown-disabled"
+            label="label"
+            items={singleMockItems}
+            placeholder="Placeholder"
+            disabledValues={["id-sre"]}
+            isOpen
+          />
+        </StoryWrapper.Row>
+      </StoryWrapper>
+    </div>
+
+    <StoryWrapper isReversed={isReversed}>
+      <StoryWrapper.Row rowTitle="Dropdown Fullwidth">
+        <Select
+          id="select-dropdown-fullwidth"
+          label="label"
+          items={singleMockItems}
+          description="This is a description"
+          placeholder="Placeholder"
+          isOpen
+          isFullWidth
+        />
+      </StoryWrapper.Row>
+    </StoryWrapper>
+  </>
+)
+
+export const DropdownStickerSheet = DropdownSheet.bind({})
+DropdownStickerSheet.storyName = "Sticker Sheet Dropdown"
+DropdownStickerSheet.parameters = {
   chromatic: { disable: false },
   controls: { disable: true },
 }