+++ packages/select/src/Select/components/TriggerButton/TriggerButton.tsx
@@ -1,6 +1,5 @@
 import React, { HTMLAttributes } from "react"
 import { AriaButtonProps, useButton } from "@react-aria/button"
-import { useFocusRing } from "@react-aria/focus"
 import { mergeProps, useObjectRef } from "@react-aria/utils"
 import { DOMAttributes, FocusableElement } from "@react-types/shared"
 import classnames from "classnames"
@@ -10,10 +9,8 @@ import chevronDown from "@kaizen/component-library/icons/chevron-down.icon.svg"
 import chevronUp from "@kaizen/component-library/icons/chevron-up.icon.svg"
 import { useSelectContext } from "../../context"
 import styles from "./TriggerButton.module.scss"
-
-export type TriggerButtonProps = OverrideClassName<
-  HTMLAttributes<HTMLButtonElement>
-> & {
+export interface TriggerButtonProps
+  extends OverrideClassName<HTMLAttributes<HTMLButtonElement>> {
   placeholder?: string
   triggerProps: AriaButtonProps<"button">
   valueProps: DOMAttributes<FocusableElement>
@@ -31,6 +28,7 @@ export const TriggerButton = React.forwardRef<
       triggerProps,
       valueProps,
       status,
+      ...restProps
     },
     buttonRef
   ) => {
@@ -38,24 +36,23 @@ export const TriggerButton = React.forwardRef<
     const value = state?.selectedItem?.rendered
     const ref = useObjectRef(buttonRef)
     const { buttonProps } = useButton(triggerProps, ref)
-    const { isFocusVisible, focusProps } = useFocusRing()
 
     return (
       <button
-        {...mergeProps(buttonProps, focusProps)}
+        {...mergeProps(buttonProps, restProps)}
         ref={ref}
         className={classnames([
           styles.button,
           (value === null || value === undefined) && styles.placeholder,
-          isFocusVisible && styles.isFocusVisible,
           status === "error" && styles.error,
           status === "caution" && styles.caution,
           triggerProps.isDisabled && styles.disabled,
           classNameOverride,
         ])}
-        disabled={triggerProps.isDisabled}
       >
-        <span {...valueProps}>{value ?? placeholder}</span>
+        <span {...valueProps} className={styles.value}>
+          {value ?? placeholder}
+        </span>
         <Icon
           icon={state.isOpen ? chevronUp : chevronDown}
           role="presentation"