+++ draft-packages/badge/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-badge@1.13.4...@kaizen/draft-badge@1.13.5) (2023-01-18)
+
+**Note:** Version bump only for package @kaizen/draft-badge
+
+
+
+
+
 ## [1.13.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-badge@1.13.3...@kaizen/draft-badge@1.13.4) (2023-01-04)
 
 **Note:** Version bump only for package @kaizen/draft-badge