+++ draft-packages/guidance-block/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.1.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@6.1.4...@kaizen/draft-guidance-block@6.1.5) (2023-01-18)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
 ## [6.1.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@6.1.3...@kaizen/draft-guidance-block@6.1.4) (2023-01-04)
 
 **Note:** Version bump only for package @kaizen/draft-guidance-block