+++ packages/tailwind/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.2.6...@kaizen/tailwind@0.2.7) (2023-01-18)
+
+**Note:** Version bump only for package @kaizen/tailwind
+
+
+
+
+
+## [0.2.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.2.5...@kaizen/tailwind@0.2.6) (2023-01-18)
+
+**Note:** Version bump only for package @kaizen/tailwind
+
+
+
+
+
+## [0.2.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.2.4...@kaizen/tailwind@0.2.5) (2023-01-18)
+
+**Note:** Version bump only for package @kaizen/tailwind
+
+
+
+
+
 ## [0.2.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.2.3...@kaizen/tailwind@0.2.4) (2023-01-05)
 
 **Note:** Version bump only for package @kaizen/tailwind